refactor out useAddDraftLineKeyBinding hook
